DocumentWordExtractor/requirements.txt
PySide6>=6.6.0
requests>=2.28.0
openpyxl>=3.1.0
python-docx>=1.0.0
PyMuPDF>=1.24.0
